% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\name{d.inv.logit}
\alias{d.inv.logit}
\title{Derivative of inverse logit}
\usage{
d.inv.logit(x)
}
\arguments{
\item{x}{}
}
\value{
vector of derivative values
}
\description{
Derivative of inverse logit
}
