% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dr.R
\name{fit.dr}
\alias{fit.dr}
\title{Fit oracle doubly robust estimating equation}
\usage{
fit.dr(
  data,
  args,
  a,
  start = NULL,
  return.var = TRUE,
  mean.a = NULL,
  cov.a = NULL,
  coef.a.l = NULL,
  var.a.l = NULL
)
}
\arguments{
\item{data}{a data frame including coluns
\itemize{
\item{outcome Y}
\item{exposures A1, ..., Am}
\item{covariates L1, ..., Lp}
}}

\item{args}{a list of arguments including
\itemize{
\item{\code{inv.link}: a function, inverse link function}
\item{\code{d.inv.link}: a function, derivative of inv.link}
\item{\code{formula}: a character string of outcome model formula}
}}

\item{start}{an optional numeric vector, starting parameter values}

\item{return.var}{an indicator for whether empirical sandwich variance
estimator should be computed, default is TRUE}

\item{mean.a}{a numeric vector, the marginal mean of the exposure A}

\item{cov.a}{a numeric matrix, the marginal covariance of the exposure A}
}
\value{
a list of arguments including
\itemize{
\item{\code{est}: root of estimating function}
\item{\code{var}: estimated covariance matrix of estimator (if requested)}
}
}
\description{
Fit oracle doubly robust estimating equation
}
