% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_equations.R
\name{fit.gfmla.mccs}
\alias{fit.gfmla.mccs}
\title{Fit MCCS G-formula estimating equation}
\usage{
fit.gfmla.mccs(
  Y,
  Astar,
  L,
  a,
  formula,
  var.e,
  B,
  seed = 123,
  inv.link,
  d.inv.link,
  start = NULL,
  return.var = TRUE
)
}
\arguments{
\item{Y}{a numeric vector, outcome variable}

\item{Astar}{a matrix of numbers, mismeasured continuous exposure}

\item{L}{a vector of numbers, confounding variable}

\item{a}{exposure value at which to estimate mean potential outcome}

\item{var.e}{a numeric vector, measurement error variance}

\item{B}{a non-negative integer, number of Monte-Carlo replicates}

\item{seed}{a non-negative integer, random seed for Monte-Carlo simulation}

\item{inv.link}{a function, inverse link function}

\item{d.inv.link}{a function, derivative of inv.link}

\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of GLM estimating function and estimated E{Y(a)}
}
\description{
Fit MCCS G-formula estimating equation
}
