% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit.glm}
\alias{fit.glm}
\title{Fit oracle GLM estimating equation}
\usage{
fit.glm(data, args, start = NULL, return.var = TRUE)
}
\arguments{
\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of GLM estimating function
}
\description{
Fit oracle GLM estimating equation
}
