% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ipw.R
\name{fit.ipw.mccs}
\alias{fit.ipw.mccs}
\title{Fit MCCS IPW estimating equation}
\usage{
fit.ipw.mccs(
  data,
  args,
  cov.e,
  B,
  mc.seed = 123,
  start = NULL,
  return.var = TRUE,
  mean.a = NULL,
  cov.a = NULL,
  coef.a.l = NULL,
  var.a.l = NULL
)
}
\arguments{
\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of MCCS IPW estimating function
}
\description{
Fit MCCS IPW estimating equation
}
