% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{get.psi.glm.mccs.}
\alias{get.psi.glm.mccs.}
\title{MCCS GLM estimating function}
\usage{
get.psi.glm.mccs.(
  data,
  g,
  formula,
  inv.link,
  d.inv.link,
  var.e,
  B,
  mc.seed = 123,
  return.sums = T
)
}
\arguments{
\item{g}{a numeric vector, coefficients in GLM Y~A+L}

\item{inv.link}{a function, inverse link function}

\item{d.inv.link}{a function, derivative of inv.link}

\item{var.e}{a numeric vector, measurement error variance}

\item{B}{a non-negative integer, number of Monte-Carlo replicates}

\item{return.sums}{a logical, indicator for whether a sum of estimating
function values (as opposed to individual values) is to be returned
(default is TRUE)}

\item{Astar}{a matrix of numbers, mismeasured continuous exposure}

\item{seed}{a non-negative integer, random seed for Monte-Carlo simulation}
}
\value{
individual or summation estimating function values
}
\description{
MCCS GLM estimating function
}
