% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{get.psi.ipw.mccs}
\alias{get.psi.ipw.mccs}
\title{MCCS IPW estimating function}
\usage{
get.psi.ipw.mccs(
  Y,
  Astar,
  L,
  g,
  inv.link,
  d.inv.link,
  var.e,
  B = 10,
  seed = 123,
  return.sums = T,
  coef.a.l,
  var.a.l,
  mean.a,
  cov.a
)
}
\arguments{
\item{Y}{a numeric vector, outcome variable}

\item{Astar}{a matrix of numbers, mismeasured continuous exposure}

\item{L}{a vector of numbers, confounding variable}

\item{g}{a numeric vector, coefficients in the MSM Y~A}

\item{inv.link}{a function, link function}

\item{d.inv.link}{a function, derivative of inv.link}

\item{var.e}{a numeric vector, measurement error variance}

\item{B}{a non-negative integer, number of Monte-Carlo replicates}

\item{seed}{a non-negative integer, random seed for Monte-Carlo simulation}

\item{return.sums}{a logical, indicator for whether a sum of estimating
function values (as opposed to individual values) is to be returned
(default is TRUE)}

\item{coef.a.l}{a vector of numbers, coefficients for the linear model A~L}

\item{var.a.l}{a vector of numbers, values of the covariance matrix of A|L}

\item{mean.a}{a vector of numbers, the marginal mean of the exposure A}

\item{cov.a}{a matrix of numbers, the marginal covariance of the exposure A}
}
\description{
MCCS IPW estimating function
}
