% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{get.psi.ps.num}
\alias{get.psi.ps.num}
\title{Propensity score numerator estimating function}
\usage{
get.psi.ps.num(data, cov.a, mean.a, return.sums = T)
}
\arguments{
\item{data}{a data frame including columns
\itemize{
\item{outcome Y}
\item{exposures A1, ..., Am}
\item{covariates L1, ..., Lp}
\item{case-cohort sampling weights cc.wts (optional)}
}}

\item{cov.a}{a numeric matrix, covariance of A}

\item{mean.a}{a numeric vector, mean of A}

\item{return.sums}{a logical, indicator for whether a sum of estimating
function values (as opposed to individual values) is to be returned
(default is TRUE)}
}
\value{
individual or summation estimating function values
}
\description{
Propensity score numerator estimating function
}
