% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{get.psi.ps}
\alias{get.psi.ps}
\title{Propensity score estimating function}
\usage{
get.psi.ps(data, ps.formula, coef.a.l, var.a.l, return.sums = T)
}
\arguments{
\item{data}{a data frame including columns
\itemize{
\item{outcome Y}
\item{exposures A1, ..., Am}
\item{covariates L1, ..., Lp}
\item{case-cohort sampling weights cc.wts (optional)}
}}

\item{ps.formula}{a character string of propensity score model formula}

\item{coef.a.l}{a numeric matrix, coefficients in propensity score model}

\item{var.a.l}{a numeric vector, variance of A|L in propensity score model}

\item{return.sums}{a logical, indicator for whether a sum of estimating
function values (as opposed to individual values) is to be returned
(default is TRUE)}
}
\value{
individual or summation estimating function values
}
\description{
Propensity score estimating function
}
