% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich_estimation.R
\name{get.sand.est.bc}
\alias{get.sand.est.bc}
\title{Bias-corrected sandwich variance estimator}
\usage{
get.sand.est.bc(ghat, get.psi, n, b = 0.7)
}
\arguments{
\item{ghat}{a numeric vector, estimated parameters}

\item{get.psi}{estimating function}

\item{n}{a positive integer, the sample size}

\item{b}{bias correction parameter}
}
\value{
estimated covariance matrix of ghat
}
\description{
Bias-corrected sandwich variance estimator
}
