% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{get.SW}
\alias{get.SW}
\title{Compute standardized IP weights for multivariate normal exposure}
\usage{
get.SW(data, ps.formula, coef.a.l, var.a.l, mean.a, cov.a)
}
\arguments{
\item{data}{a data frame including columns
\itemize{
\item{outcome Y}
\item{exposures A1, ..., Am}
\item{covariates L1, ..., Lp}
\item{case-cohort sampling weights cc.wts (optional)}
}}

\item{ps.formula}{a character string of propensity score model formula}

\item{coef.a.l}{a numeric matrix, coefficients in propensity score model}

\item{var.a.l}{a numeric vector, variance of A|L in propensity score model}

\item{mean.a}{a numeric vector, the marginal mean of the exposure A}

\item{cov.a}{a numeric matrix, the marginal covariance of the exposure A}
}
\value{
a vector of standardized IP weights
}
\description{
Compute standardized IP weights for multivariate normal exposure
}
