% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{make.mccs}
\alias{make.mccs}
\title{create MCCS estimating function}
\usage{
make.mccs(get.psi, data, args, cov.e, B, mc.seed)
}
\arguments{
\item{psi}{an estimating function}
}
\value{
an MCCS estimating function
}
\description{
create MCCS estimating function
}
