% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dr.R
\name{sim.dr}
\alias{sim.dr}
\title{run one DR simulation}
\usage{
sim.dr(n, vare, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{B}{a positive integer, the number of MC replicates}

\item{seed}{a positive integer, random number seed set at start of simulation}
}
\value{
a data frame with the following columns
\itemize{
\item{n}
\item{B}
\item{vare}
\item{seed}
\item{ps: an indicator for whether the propensity score model is correctly
(0) or incorrectly (1) specified}
\item{out: an indicator for whether the outcome model is correctly
(0) or incorrectly (1) specified}
\item{type: the type of estimator (naive, oracle, or mccs)}
\item{est: estimated slope of MSM}
\item{se: estimated standard error}
}
}
\description{
run one DR simulation
}
