% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gfmla_2.R
\name{sim.gfmla.2}
\alias{sim.gfmla.2}
\title{run one G-formula 2 simulation}
\usage{
sim.gfmla.2(n, a, vare, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{a}{exposure value at which to estimate mean potential outcome}

\item{vare}{a non-negative number, measurement error variance}

\item{B}{a positive integer, the number of MC replicates}

\item{seed}{a positive integer, random number seed set at start of simulation}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{a1, ..., ak}
\item{est.OG: oracle g-formula estinates}
\item{est.NG: naive g-formula estinates}
\item{est.CG: corrected g-formula estinates}
}
}
\description{
run one G-formula 2 simulation
}
