% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ipw_nonadd.R
\name{sim.ipw.nonadd}
\alias{sim.ipw.nonadd}
\title{run one IPW simulation with a trivariate exposure, binary outcome, a linear
link function, and non-additive measurement error}
\usage{
sim.ipw.nonadd(n, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{B}{a positive integer, the number of MC replicates}

\item{seed}{a positive integer, random number seed set at start of simulation}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{ghat.OL: oracle linear regression estinates}
\item{ghat.NL: naive linear regression estinates}
\item{ghat.CL: corrected linear regression estinates}
\item{ghat.OG: oracle IPW estinates}
\item{ghat.NG: naive IPW estinates}
\item{ghat.CG: corrected IPW estinates}
\item{stde.OL: oracle linear regression standard errors}
\item{stde.NL: naive linear regression standard errors}
\item{stde.CL: corrected linear regression standard errors}
\item{stde.OG: oracle IPW standard errors}
\item{stde.NG: naive IPW standard errors}
\item{stde.CG: corrected IPW standard errors}
}
}
\description{
run one IPW simulation with a trivariate exposure, binary outcome, a linear
link function, and non-additive measurement error
}
