% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1_gfmla_nonlinear_coarse.R
\name{sim1.gfmla.nonlinear.coarse}
\alias{sim1.gfmla.nonlinear.coarse}
\title{Simulation 1: g-formula with nonlinear MSM and coarse grid}
\usage{
sim1.gfmla.nonlinear.coarse(n, a, vare, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{a}{exposure value at which to estimate mean potential outcome}

\item{vare}{a non-negative number, the measurement error variance}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{seed}{a positive integer, the random number seed}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{a1, ..., ak}
\item{est.xx: estimates of dose response curve using various methods}
\item{ste.xx: standard errors using various methods}
\item{bcs.xx: bias-corrected standard errors using various methods}
}
}
\description{
Simulation 1: g-formula with nonlinear MSM and coarse grid
}
