% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim10_dr_cc.R
\name{sim10.dr.cc}
\alias{sim10.dr.cc}
\title{Simulation 10: DR with case cohort sampling}
\usage{
sim10.dr.cc(n, vare, B, pi.cc, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{vare}{a non-negative number, the measurement error variance}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{pi.cc}{a number in (0,1], the case-cohort sampling proportion}

\item{seed}{a positive integer, the random number seed}
}
\value{
a data frame with the following columns
\itemize{
\item{n}
\item{B}
\item{vare}
\item{pi.cc}
\item{seed}
\item{ps: an indicator for whether the propensity score model is correctly
(0) or incorrectly (1) specified}
\item{out: an indicator for whether the outcome model is correctly
(0) or incorrectly (1) specified}
\item{type: the type of estimator (naive, oracle, reg. cal., SIMEX, or MCCS)}
\item{est: estimated MSM parameter}
\item{se: standard error}
\item{bse: bias-corrected standard error}
}
}
\description{
Simulation 10: DR with case cohort sampling
}
