% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2_gfmla_nonlinear_fine.R
\name{sim2.gfmla.nonlinear.fine}
\alias{sim2.gfmla.nonlinear.fine}
\title{Simulation 2: g-formula with nonlinear MSM and coarse grid}
\usage{
sim2.gfmla.nonlinear.fine(n, a, vare, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{a}{exposure value at which to estimate mean potential outcome}

\item{vare}{a non-negative number, the measurement error variance}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{seed}{a positive integer, the random number seed}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{a1, ..., ak}
\item{est.xx: estimates of dose response curve using various methods}
}
}
\description{
Simulation 2: g-formula with nonlinear MSM and coarse grid
}
