% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim3_ipw_nonlinearps.R
\name{sim3.ipw.nonlinearps}
\alias{sim3.ipw.nonlinearps}
\title{Simulation 3: IPW with nonlinear PS model}
\usage{
sim3.ipw.nonlinearps(n, vare, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{vare}{a non-negative number, the measurement error variance}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{seed}{a positive integer, the random number seed}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{ghat.xx: MSM parameter estimates for various methods}
\item{stde.xx: standard errors for various methods}
\item{bste.xx: bias-corrected standard errors for various methods}
}
}
\description{
Simulation 3: IPW with nonlinear PS model
}
