% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim4_dr_nonlinearps.R
\name{sim4.dr.nonlinearps}
\alias{sim4.dr.nonlinearps}
\title{Simulation 4: DR with nonlinear PS model}
\usage{
sim4.dr.nonlinearps(n, vare, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{vare}{a non-negative number, the measurement error variance}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{seed}{a positive integer, the random number seed}
}
\value{
a data frame with the following columns
\itemize{
\item{n}
\item{B}
\item{vare}
\item{seed}
\item{ps: an indicator for whether the propensity score model is correctly
(0) or incorrectly (1) specified}
\item{out: an indicator for whether the outcome model is correctly
(0) or incorrectly (1) specified}
\item{type: the type of estimator (naive, oracle, reg. cal., SIMEX, or MCCS)}
\item{est: estimated MSM parameter}
\item{se: standard error}
\item{bse: bias-corrected standard error}
}
}
\description{
Simulation 4: DR with nonlinear PS model
}
