% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim5_gfmla_reliability.R
\name{sim5.gfmla.reliability}
\alias{sim5.gfmla.reliability}
\title{Simulation 5: g-formula with varying reliability}
\usage{
sim5.gfmla.reliability(n, a, rel, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{a}{exposure value at which to estimate mean potential outcome}

\item{rel}{a number in \link{0,1}, the measurement error reliability}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{seed}{a positive integer, the random number seed}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{a1, ..., ak}
\item{est.xx: estimates of dose response curve using various methods}
\item{ste.xx: standard errors using various methods}
\item{bcs.xx: bias-corrected standard errors using various methods}
}
}
\description{
Simulation 5: g-formula with varying reliability
}
