% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim6_ipw_reliability.R
\name{sim6.ipw.reliability}
\alias{sim6.ipw.reliability}
\title{Simulation 6: IPW with varying reliability}
\usage{
sim6.ipw.reliability(n, rel, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{seed}{a positive integer, the random number seed}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{ghat.xx: MSM parameter estimates for various methods}
\item{stde.xx: standard errors for various methods}
\item{bste.xx: bias-corrected standard errors for various methods}
}
}
\description{
Simulation 6: IPW with varying reliability
}
