% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim7_ipw_estvar.R
\name{sim7.ipw.estvar}
\alias{sim7.ipw.estvar}
\title{Simulation 7: IPW with estimated measurement error variance}
\usage{
sim7.ipw.estvar(n, vare, B, n.supp, k, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{vare}{a non-negative number, the measurement error variance}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{n.supp}{number of observations with replicate exposure measurements}

\item{k}{number of measurement error replicates}

\item{seed}{a positive integer, the random number seed}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{ghat.xx: MSM parameter estimates for various methods}
\item{stde.xx: standard errors for various methods}
\item{bste.xx: bias-corrected standard errors for various methods}
}
}
\description{
Simulation 7: IPW with estimated measurement error variance
}
