% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim8_ipw_multiplicative.R
\name{sim8.ipw.multiplicative}
\alias{sim8.ipw.multiplicative}
\title{Simulation 8: IPW with multiplicative measurement error}
\usage{
sim8.ipw.multiplicative(n, rel, n.supp, k, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{seed}{a positive integer, the random number seed}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{ghat.xx: MSM parameter estimates for various methods}
\item{stde.xx: standard errors for various methods}
\item{bste.xx: bias-corrected standard errors for various methods}
}
}
\description{
Simulation 8: IPW with multiplicative measurement error
}
