% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim9_ipw_positivity.R
\name{sim9.ipw.positivity}
\alias{sim9.ipw.positivity}
\title{Simulation 9: IPW with positivity violation}
\usage{
sim9.ipw.positivity(n, vare, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{vare}{a non-negative number, the measurement error variance}

\item{B}{a positive integer, the number of Monte-Carlo replicates for MCCS}

\item{seed}{a positive integer, the random number seed}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{ghat.xx: MSM parameter estimates for various methods}
\item{stde.xx: standard errors for various methods}
\item{bste.xx: bias-corrected standard errors for various methods}
}
}
\description{
Simulation 9: IPW with positivity violation
}
