% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_B.R
\name{tune.B}
\alias{tune.B}
\title{Tune MCCS estimating equation to select B value}
\usage{
tune.B(get.psi, data, cov.e, BB, args = list(), mc.seed = 123)
}
\description{
Tune MCCS estimating equation to select B value
}
