% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dr.R
\name{fit.dr}
\alias{fit.dr}
\title{Fit oracle doubly robust estimating equation}
\usage{
fit.dr(
  Y,
  A,
  L,
  a,
  formula,
  inv.link,
  d.inv.link,
  start = NULL,
  return.var = TRUE,
  mean.a = NULL,
  cov.a = NULL,
  coef.a.l = NULL,
  var.a.l = NULL
)
}
\arguments{
\item{Y}{a numeric vector, outcome variable}

\item{A}{a matrix of numbers, possibly multivariate continuous exposure
values with rows corresponding to observations}

\item{L}{a vector of numbers, confounding variable}

\item{a}{exposure value at which to estimate mean potential outcome}

\item{inv.link}{a function, inverse link function}

\item{d.inv.link}{a function, derivative of inv.link}

\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of GLM estimating function
}
\description{
Fit oracle doubly robust estimating equation
}
