% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gfmla.R
\name{fit.gfmla.mccs}
\alias{fit.gfmla.mccs}
\title{Fit MCCS G-formula estimating equation}
\usage{
fit.gfmla.mccs(
  data,
  args,
  a,
  cov.e,
  B,
  mc.seed,
  start = NULL,
  return.var = TRUE
)
}
\arguments{
\item{a}{exposure value at which to estimate mean potential outcome}

\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of GLM estimating function and estimated E{Y(a)}
}
\description{
Fit MCCS G-formula estimating equation
}
