% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gfmla.R
\name{fit.gfmla.mccs}
\alias{fit.gfmla.mccs}
\title{Fit MCCS G-formula estimating equation}
\usage{
fit.gfmla.mccs(
  Y,
  Astar,
  L,
  a,
  formula,
  var.e,
  B,
  seed = 123,
  inv.link,
  d.inv.link,
  start = NULL,
  return.var = TRUE
)
}
\arguments{
\item{a}{exposure value at which to estimate mean potential outcome}

\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of GLM estimating function and estimated E{Y(a)}
}
\description{
Fit MCCS G-formula estimating equation
}
