% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gfmla.R
\name{fit.gfmla}
\alias{fit.gfmla}
\title{Fit oracle G-formula estimating equation}
\usage{
fit.gfmla(data, args, a, start = NULL, return.var = TRUE)
}
\arguments{
\item{a}{exposure value at which to estimate mean potential outcome}

\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of GLM estimating function
}
\description{
Fit oracle G-formula estimating equation
}
