% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit.glm.mccs}
\alias{fit.glm.mccs}
\title{Fit MCCS GLM estimating equation}
\usage{
fit.glm.mccs(data, args, cov.e, B, mc.seed, start = NULL, return.var = TRUE)
}
\arguments{
\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of MCCS GLM estimating function
}
\description{
Fit MCCS GLM estimating equation
}
