% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_equations.R
\name{fit.glm.mccs}
\alias{fit.glm.mccs}
\title{Fit MCCS GLM estimating equation}
\usage{
fit.glm.mccs(
  Y,
  Astar,
  L,
  var.e,
  B = 10,
  seed = 123,
  inv.link,
  d.inv.link,
  start = NULL
)
}
\arguments{
\item{Y}{a numeric vector, outcome variable}

\item{Astar}{a matrix of numbers, mismeasured continuous exposure}

\item{L}{a vector of numbers, confounding variable}

\item{var.e}{a numeric vector, measurement error variance}

\item{B}{a non-negative integer, number of Monte-Carlo replicates}

\item{seed}{a non-negative integer, random seed for Monte-Carlo simulation}

\item{inv.link}{a function, inverse link function}

\item{d.inv.link}{a function, derivative of inv.link}

\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of MCCS GLM estimating function
}
\description{
Fit MCCS GLM estimating equation
}
