% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_equations.R
\name{fit.glm}
\alias{fit.glm}
\title{Fit oracle GLM estimating equation}
\usage{
fit.glm(
  Y,
  A,
  L,
  formula,
  inv.link,
  d.inv.link,
  start = NULL,
  return.var = TRUE
)
}
\arguments{
\item{Y}{a numeric vector, outcome variable}

\item{A}{a matrix of numbers, possibly multivariate continuous exposure
values with rows corresponding to observations}

\item{L}{a vector of numbers, confounding variable}

\item{inv.link}{a function, inverse link function}

\item{d.inv.link}{a function, derivative of inv.link}

\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of GLM estimating function
}
\description{
Fit oracle GLM estimating equation
}
