% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit.glm}
\alias{fit.glm}
\title{Fit oracle GLM estimating equation}
\usage{
fit.glm(data, args, start = NULL, return.var = TRUE, return.bcvar = TRUE)
}
\arguments{
\item{data}{a data frame including columns
\itemize{
\item{outcome Y}
\item{exposures A1, ..., Am}
\item{covariates L1, ..., Lp}
\item{case-cohort sampling weights cc.wts (optional)}
}}

\item{args}{a list of arguments including
\itemize{
\item{\code{inv.link}: a function, inverse link function}
\item{\code{d.inv.link}: a function, derivative of inv.link}
\item{\code{formula}: a character string of outcome model formula}
}}

\item{start}{an optional numeric vector, starting parameter values}

\item{return.var}{an indicator for whether empirical sandwich variance
estimator should be computed, default is TRUE}

\item{return.bcvar}{an indicator for whether bias-corrected variance
estimator should be computed, default is TRUE}
}
\value{
a list of arguments including
\itemize{
\item{\code{est}: root of estimating function}
\item{\code{var}: estimated covariance matrix of estimator (if requested)}
\item{\code{bc.var}: bias-corrected estimated covariance matrix of estimator
(if requested)}
}
}
\description{
Fit oracle GLM estimating equation
}
