% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ipw.R
\name{fit.ipw.mccs}
\alias{fit.ipw.mccs}
\title{Fit MCCS IPW estimating equation}
\usage{
fit.ipw.mccs(
  Y,
  Astar,
  L,
  formula,
  ps.formula,
  var.e,
  B,
  seed = 123,
  inv.link,
  d.inv.link,
  start = NULL,
  return.var = TRUE,
  mean.a = NULL,
  cov.a = NULL,
  coef.a.l = NULL,
  var.a.l = NULL
)
}
\arguments{
\item{Y}{a numeric vector, outcome variable}

\item{L}{a vector of numbers, confounding variable}

\item{inv.link}{a function, link function}

\item{d.inv.link}{a function, derivative of inv.link}

\item{start}{an optional numeric vector, starting parameter values}

\item{coef.a.l}{a vector of numbers, coefficients for the linear model A~L}

\item{var.a.l}{a vector of numbers, values of the covariance matrix of A|L}
}
\value{
root of MCCS IPW estimating function
}
\description{
Fit MCCS IPW estimating equation
}
