% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_equations.R
\name{fit.ipw}
\alias{fit.ipw}
\title{Fit oracle IPW estimating equation}
\usage{
fit.ipw(
  Y,
  A,
  L,
  inv.link,
  d.inv.link,
  start = NULL,
  coef.a.l = NULL,
  var.a.l = NULL
)
}
\arguments{
\item{Y}{a numeric vector, outcome variable}

\item{A}{a matrix of numbers, possibly multivariate continuous exposure
values with rows corresponding to observations}

\item{L}{a vector of numbers, confounding variable}

\item{inv.link}{a function, link function}

\item{d.inv.link}{a function, derivative of inv.link}

\item{start}{an optional numeric vector, starting parameter values}

\item{coef.a.l}{a vector of numbers, coefficients for the linear model A~L}

\item{var.a.l}{a vector of numbers, values of the covariance matrix of A|L}
}
\value{
root of IPW estimating function
}
\description{
Fit oracle IPW estimating equation
}
