% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ipw.R
\name{fit.ipw}
\alias{fit.ipw}
\title{Fit oracle IPW estimating equation}
\usage{
fit.ipw(
  data,
  args,
  start = NULL,
  return.var = TRUE,
  mean.a = NULL,
  cov.a = NULL,
  coef.a.l = NULL,
  var.a.l = NULL
)
}
\arguments{
\item{start}{an optional numeric vector, starting parameter values}
}
\value{
root of IPW estimating function
}
\description{
Fit oracle IPW estimating equation
}
