% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{get.psi.ipw}
\alias{get.psi.ipw}
\title{Oracle IPW estimating function}
\usage{
get.psi.ipw(data, g, args, mean.a, cov.a, return.sums = T)
}
\arguments{
\item{g}{a numeric vector, coefficients in the MSM Y~A}

\item{return.sums}{a logical, indicator for whether a sum of estimating
function values (as opposed to individual values) is to be returned
(default is TRUE)}

\item{Y}{a numeric vector, outcome variable}

\item{inv.link}{a function, link function}

\item{d.inv.link}{a function, derivative of inv.link}
}
\value{
individual or summation estimating function values
}
\description{
Oracle IPW estimating function
}
