% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{get.psi.ps}
\alias{get.psi.ps}
\title{propensity score estimating function}
\usage{
get.psi.ps(A, L, ps.formula, coef.a.l, var.a.l, return.sums = T)
}
\arguments{
\item{A}{a matrix of numbers, possibly multivariate continuous exposure
values with rows corresponding to observations}

\item{L}{a vector of numbers, confounding variable}

\item{coef.a.l}{a vector of numbers, coefficients for the linear model A~L}

\item{var.a.l}{a vector of numbers, values of the covariance matrix of A|L}

\item{return.sums}{a logical, indicator for whether a sum of estimating
function values (as opposed to individual values) is to be returned
(default is TRUE)}
}
\value{
individual or summation estimating function values
}
\description{
propensity score estimating function
}
