% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{get.SW}
\alias{get.SW}
\title{compute standardized IP weights for multivariate normal exposure}
\usage{
get.SW(A, L, coef.a.l, var.a.l, mean.a, cov.a)
}
\arguments{
\item{A}{a matrix of numbers, possibly multivariate continuous exposure
values with rows corresponding to observations}

\item{L}{a vector of numbers, confounding variable}

\item{coef.a.l}{a vector of numbers, coefficients for the linear model A~L}

\item{var.a.l}{a vector of numbers, values of the covariance matrix of A|L}

\item{mean.a}{a vector of numbers, the marginal mean of the exposure A}

\item{cov.a}{a matrix of numbers, the marginal covariance of the exposure A}
}
\value{
a vector of standardized IP weights
}
\description{
compute standardized IP weights for multivariate normal exposure
}
