% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_functions.R
\name{make.mccs}
\alias{make.mccs}
\title{create MCCS estimating function}
\usage{
make.mccs(get.psi, data, args, cov.e, B, mc.seed)
}
\arguments{
\item{get.psi}{an estimating function}

\item{data}{a data frame including columns
\itemize{
\item{outcome Y}
\item{exposures A1, ..., Am}
\item{covariates L1, ..., Lp}
\item{case-cohort sampling weights cc.wts (optional)}
}}

\item{args}{a list of additional arguments}

\item{cov.e}{a numeric matrix, measurement error covariance}

\item{B}{a positive integer, the number of MC replicates}

\item{mc.seed}{a positive integer, random number seed for MC simulation}
}
\value{
an MCCS estimating function
}
\description{
create MCCS estimating function
}
