% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{mod.mat}
\alias{mod.mat}
\title{create model matrix with complex numbers}
\usage{
mod.mat(trms, data, contrasts.arg = NULL)
}
\arguments{
\item{trms}{a terms object}
}
\value{
model matrix
}
\description{
create model matrix with complex numbers
}
