% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim.ipw}
\alias{sim.ipw}
\title{run one IPW simulation with a trivariate exposure, binary outcome, and a linear link function}
\usage{
sim.ipw(n, vare, B, seed)
}
\arguments{
\item{n}{a positive integer, the sample size}

\item{vare}{a non-negative number, the measurement error variance for the first component of the exposure}

\item{B}{a non-negative integer, the number of Monte-Carlo replicates used in corrected score methods}

\item{seed}{a non-negative integer, the random number seed to be set before data are generated}
}
\value{
a named numeric vector with the following entries
\itemize{
\item{n}
\item{vare}
\item{B}
\item{seed}
\item{ghat: the estimated marginal structural model parameters from six methods (OL = oracle linear model, NL = naive linear model, CL = corrected linear model, OI = oracle IPW, NI = naive IPW, CI = corrected IPW)}
\item{evar: the estimated variance of ghat for each of the six methods}
}
}
\description{
run one IPW simulation with a trivariate exposure, binary outcome, and a linear link function
}
